docs: add docs/ folder, timezone config note, clean up UI lint
- New docs/ with one page per high-level feature: rooms, cycle engine,
  vent control methods, thermostat settings, schedules, presence,
  system modes, observability, backup/restore, MCP server. Each page
  is grounded in what the code actually does.
- README links to docs/ from a new "Documentation" section.
- README has a "Timezone configuration" section directing users to set
  the timezone via the add-on Configuration tab in Home Assistant —
  the default UTC misfires schedules otherwise.
- UI lint is now clean (0 warnings):
  - Split SystemContext / DevModeContext and their hooks into
    src/contexts.ts so they're no longer mixed with component exports.
  - Moved the app shell (AppRoot, Nav, SystemToggle, DevModeToggle)
    into a new src/App.tsx so src/main.tsx is purely a boot file.
    Fixes react-refresh warnings on the entry file.
  - Added targeted eslint-disable comments with explanations on the
    useEffect hooks whose dependency lists are intentionally scoped
    (filter-only reloads, mount-only effects using refs).
