+++ smart_vent/frontend/src/pages/Rooms.tsx
@@ -24,7 +24,7 @@ import {
   type EntityState,
   type RoomActiveStatus,
 } from "../api";
-import { useSystem } from "../main";
+import { useSystem } from "../contexts";
 import EntityPicker from "../components/EntityPicker";
 
 // ---------------------------------------------------------------------------
@@ -718,6 +718,9 @@ function RoomCard({
     getEntityStates(allIds)
       .then(setStates)
       .catch(() => {});
+    // Intentionally scoped to room.id: refetching on every sensor/vent/presence
+    // mutation would thrash HA during inline edits.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [room.id]);
 
   // Derived live values
@@ -925,9 +928,12 @@ export default function Rooms() {
 
   useEffect(() => {
     load();
+    // Mount-only.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  // Re-fetch statuses every 30s
+  // Re-fetch statuses every 30s. Mount-only — uses ref to read latest rooms
+  // without re-subscribing.
   useEffect(() => {
     const interval = setInterval(() => fetchStatuses(roomsRef.current), 30_000);
     return () => clearInterval(interval);
